Transformation: rename endpoint and request model to pipelineExecution~
